# inbound-controller-ws/src/main/java/sk/posam/fsa/dto/NotificationDto.java
package sk.posam.fsa.dto;

public class NotificationDto {
    private Long id;
    private Long userId;
    private String message;
    private boolean read;

    // Gettery a settery
    public Long getId() { return id; }
    public void setId(Long id) { this.id = id; }

    public Long getUserId() { return userId; }
    public void setUserId(Long userId) { this.userId = userId; }

    public String getMessage() { return message; }
    public void setMessage(String message) { this.message = message; }

    public boolean isRead() { return read; }
    public void setRead(boolean read) { this.read = read; }
}