# inbound-controller-ws/src/main/java/sk/posam/fsa/dto/PerformanceDto.java
package sk.posam.fsa.dto;

import java.math.BigDecimal;

public class PerformanceDto {
    private Long id;
    private Long athleteId;
    private Long exerciseId;
    private BigDecimal result;

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public Long getAthleteId() {
        return athleteId;
    }

    public void setAthleteId(Long athleteId) {
        this.athleteId = athleteId;
    }

    public Long getExerciseId() {
        return exerciseId;
    }

    public void setExerciseId(Long exerciseId) {
        this.exerciseId = exerciseId;
    }

    public BigDecimal getResult() {
        return result;
    }

    public void setResult(BigDecimal result) {
        this.result = result;
    }
}